python/cugraph-nx/pyproject.toml
# Copyright (c) 2023, NVIDIA CORPORATION.

[build-system]

requires = [
    "setuptools",
    "wheel",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
build-backend = "setuptools.build_meta"

[project]
name = "cugraph-nx"
version = "23.06.00"
description = "cugraph backend for NetworkX"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "NVIDIA Corporation" },
]
license = { text = "Apache 2.0" }
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "cupy-cuda11x>=12.0.0",
    "pylibcugraph==23.6.*",
    "networkx >=3.0",
]

[project.optional-dependencies]
test = [
    "pandas",
    "pytest",
]

[project.urls]
Homepage = "https://github.com/rapidsai/cugraph"
Documentation = "https://docs.rapids.ai/api/cugraph/stable/"

[project.entry-points."networkx.plugins"]
cugraph = "cugraph_nx.interface:Dispatcher"

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
include = [
    "cugraph_nx*",
    "cugraph_nx.*",
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.isort]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
profile = "black"
skip_gitignore = true
float_to_top = true
default_section = "THIRDPARTY"
known_first_party = "cugraph_nx"
line_length = 88

[tool.pytest.ini_options]
testpaths = "cugraph_nx/tests"
xfail_strict = true
markers = [
    "slow: Skipped unless --runslow passed",
]
filterwarnings = [
    # See: https://docs.python.org/3/library/warnings.html#describing-warning-filters
    # and: https://docs.pytest.org/en/7.2.x/how-to/capture-warnings.html#controlling-warnings
    "error",
]

[tool.coverage.run]
branch = true
source = ["cugraph_nx"]
omit = []

[tool.coverage.report]
ignore_errors = false
precision = 1
fail_under = 0
skip_covered = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.ruff]
# https://github.com/charliermarsh/ruff/
line-length = 88
target-version = "py39"
select = [
    "ALL",
]
external = [
    # noqa codes that ruff doesn't know about: https://github.com/charliermarsh/ruff#external
]
ignore = [
    # Would be nice to fix these
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed in ...
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    # "D107",  # Missing docstring in `__init__`
    "D401",  # First line of docstring should be in imperative mood:
    # "D417",  # Missing argument description in the docstring:
    "PLE0605",  # Invalid format for `__all__`, must be `tuple` or `list` (Note: broken in v0.0.237)

    # Maybe consider
    # "SIM300",  # Yoda conditions are discouraged, use ... instead (Note: we're not this picky)
    # "SIM401",  # Use dict.get ... instead of if-else-block (Note: if-else better for coverage and sometimes clearer)
    # "TRY004",  # Prefer `TypeError` exception for invalid type (Note: good advice, but not worth the nuisance)
    # "TRY200",  # Use `raise from` to specify exception cause (Note: sometimes okay to raise original exception)

    # Intentionally ignored
    "A003",  # Class attribute ... is shadowing a python builtin
    "ANN101",  # Missing type annotation for `self` in method
    "ARG004",  # Unused static method argument: `...`
    "COM812",  # Trailing comma missing
    "D203",  # 1 blank line required before class docstring (Note: conflicts with D211, which is preferred)
    "D400",  # First line should end with a period (Note: prefer D415, which also allows "?" and "!")
    "F403",  # `from .classes import *` used; unable to detect undefined names (Note: used to match networkx)
    "N802",  # Function name ... should be lowercase
    "N803",  # Argument name ... should be lowercase (Maybe okay--except in tests)
    "N806",  # Variable ... in function should be lowercase
    "N807",  # Function name should not start and end with `__`
    "N818",  # Exception name ... should be named with an Error suffix (Note: good advice)
    "PLR0911",  # Too many return statements
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments to function call
    "PLR0915",  # Too many statements
    "PLR2004",  # Magic number used in comparison, consider replacing magic with a constant variable
    "PLW2901",  # Outer for loop variable ... overwritten by inner assignment target (Note: good advice, but too strict)
    "RET502",  # Do not implicitly `return None` in function able to return non-`None` value
    "RET503",  # Missing explicit `return` at the end of function able to return non-`None` value
    "RET504",  # Unnecessary variable assignment before `return` statement
    "S110",  # `try`-`except`-`pass` detected, consider logging the exception (Note: good advice, but we don't log)
    "S112",  # `try`-`except`-`continue` detected, consider logging the exception (Note: good advice, but we don't log)
    "SIM102",  # Use a single `if` statement instead of nested `if` statements (Note: often necessary)
    "SIM105",  # Use contextlib.suppress(...) instead of try-except-pass (Note: try-except-pass is much faster)
    "SIM108",  # Use ternary operator ... instead of if-else-block (Note: if-else better for coverage and sometimes clearer)
    "TRY003",  # Avoid specifying long messages outside the exception class (Note: why?)

    # Ignored categories
    "C90",  # mccabe (Too strict, but maybe we should make things less complex)
    "I",  # isort (Should we replace `isort` with this?)
    "ANN",  # flake8-annotations
    "BLE",  # flake8-blind-except (Maybe consider)
    "FBT",  # flake8-boolean-trap (Why?)
    "DJ",  # flake8-django (We don't use django)
    "EM",  # flake8-errmsg (Perhaps nicer, but too much work)
    # "ICN",  # flake8-import-conventions (Doesn't allow "_" prefix such as `_np`)
    "PYI",  # flake8-pyi (We don't have stub files yet)
    "SLF",  # flake8-self (We can use our own private variables--sheesh!)
    "TID",  # flake8-tidy-imports (Rely on isort and our own judgement)
    # "TCH",  # flake8-type-checking
    # "ARG",  # flake8-unused-arguments (Sometimes helpful, but too strict)
    "TD",  # flake8-todos (Maybe okay to add some of these)
    "ERA",  # eradicate (We like code in comments!)
    "PD",  # pandas-vet (Intended for scripts that use pandas, not libraries)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports (w/o defining `__all__`)
"cugraph_nx/**/tests/*py" = ["S101", "T201", "D103", "D100"]  # Allow assert, print, and no docstring

[tool.ruff.flake8-annotations]
mypy-init-return = true

[tool.ruff.flake8-builtins]
builtins-ignorelist = ["copyright"]

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.pydocstyle]
convention = "numpy"
